pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kairos"
version = "2.3.0"
description = "Production ML Risk Assessment System"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Abhijit More"}
]
dependencies = [
    "pandas==2.2.3",
    "numpy==1.26.4",
    "scikit-learn==1.8.0",
    "lightgbm==4.6.0",
    "catboost==1.2.8",
    "fastapi==0.128.0",
    "uvicorn==0.34.0",
    "joblib==1.4.2",
    "requests==2.32.4",
    "matplotlib==3.10.0",
    "mlflow==3.8.1",
    "optuna==4.6.0",
    "pyyaml==6.0.2",
    "prometheus-fastapi-instrumentator==7.1.0",
    "pytest==9.0.2",
    "pytest-cov==7.0.0",
    "httpx==0.27.2",
    "pydantic-settings==2.7.1",
    "slowapi==0.1.9",
    "category_encoders==2.9.0",
    "seaborn==0.13.2",
    "flask==3.0.3",
    "tqdm==4.66.4",
    "python-dotenv==1.0.1",
    "celery==5.4.0",
    "redis==5.2.1"
]

[project.optional-dependencies]
dev = [
    "mypy==1.15.0",
    "ruff==0.9.2",
    "selenium==4.21.0",
    "locust==2.28.1",
    "types-requests",
    "types-PyYAML",
    "pandas-stubs"
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["E402"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "--cov=src --cov-report=term-missing -v"
markers = [
    "unit: Small, fast tests for individual functions and classes.",
    "integration: Tests for API endpoints and component interactions.",
    "chaos: Failure injection and system robustness tests.",
    "slow: Tests that take a significant time to run.",
    "e2e: End-to-end user journey tests using browser automation."
]
